🚚: API 요청 파일 분리
