+++ src/components/List/QuestionList.jsx
@@ -5,6 +5,7 @@ import useMobileLayout from '../../utils/useMobileLayout';
 import useTabletLayout from '../../utils/useTabletLayout';
 import DropDownButton from '../Buttons/DropDownButton';
 import Pagination from './Pagination/Pagination';
+import getCardList from '../../services/fetchQuestionList';
 
 export default function QuestionList() {
   const [currentPage, setCurrentPage] = useState(1);
@@ -18,17 +19,6 @@ export default function QuestionList() {
   const isMobileSize = useMobileLayout();
   const isTabletSize = useTabletLayout();
   const LIMITSIZE = isMobileSize || isTabletSize ? MIN_CARDS : MAX_CARDS;
-
-  const getCardList = async ({
-    sort: cardSort,
-    LIMITSIZE: cardLimitSize,
-    currentPage: current,
-  }) => {
-    const url = `/subjects/?limit=${cardLimitSize}&offset=${(current - 1) * LIMITSIZE}&sort=${cardSort}`;
-    const response = await fetch(`https://openmind-api.vercel.app/3-2${url}`);
-    const result = await response.json();
-    return result;
-  };
 
   const handleLoad = async options => {
     const { results, count } = await getCardList(options);
